make LanguageClient platform-independent
Platform-specific routines remain in main.ts, but consumers can import client.ts without requiring any Node or Electron dependencies. This makes it possible to use LanguageClient in platforms other than Node+Electron.
